requirements.txt
click==8.1.7
dependency-injector==4.41.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.26.1
Pillow==10.1.0
PyYAML==6.0.1
six==1.16.0
